Workouts and Exercises models/controllers/etc. functioning; clients recent workouts functioning

+++ client/src/ClientCard.js
@@ -1,23 +1,33 @@
 import { useState } from 'react';
+import EditForm from './EditForm';
+import Workout from './Workout'
 
 function ClientCard({client, clients, setClients}){
   const[editForm, setEditForm] = useState(false)
 
   function handleEdit(){
-    console.log("edit")
+    setEditForm(!editForm)
   }
 
+  const lastThreeWorkouts = client.workouts.slice(-3)
+  const clientRecentWorkouts = lastThreeWorkouts.map((workout)=><Workout workout={workout} key={workout.id}/>)
+ 
+  function handleNextWorkout(){
+    console.log("Next Workout")
+  }
   function handleDelete(){
-    
-    fetch(`/clients/${client.id}`, {
-      method: "DELETE",
-      headers: {
-        "Content-Type": "application/json",
-      }
-    }).then(()=>{
-      const newClientList = clients.filter((c)=>c.id!==parseInt(client.id))
-      setClients(newClientList)
-    })
+    const deleteConfirm = window.confirm("Are you sure you want to delete this submission?")
+    if (deleteConfirm){
+      fetch(`/clients/${client.id}`, {
+        method: "DELETE",
+        headers: {
+          "Content-Type": "application/json",
+        }
+      }).then(()=>{
+        const newClientList = clients.filter((c)=>c.id!==parseInt(client.id))
+        setClients(newClientList)
+      })
+    }
   }
 
   return(
@@ -26,7 +36,17 @@ function ClientCard({client, clients, setClients}){
       <p>{client.email}</p>
       <p>{client.age} years, {client.feet} ft. {client.inches} in., {client.weight} lbs</p>
       <p>{client.fitness_level}, {client.workouts_per_week} workouts per week</p>
-      <button onClick={handleEdit}>Edit</button><button onClick={handleDelete}>Delete</button>
+      <button onClick={handleEdit}>{editForm ? "Close" : "Edit Client"}</button>
+      <button onClick={handleDelete}>Delete Client</button>
+      <button onClick={handleNextWorkout}>Assign Next Workout</button>
+      <div>{editForm ? <EditForm client = {client} clients = {clients} setClients = {setClients} editForm={editForm} setEditForm={setEditForm}/> : ""}</div>
+      <div className="client-workouts">
+        <p>Most recent workouts:</p>
+        <ul>
+          {clientRecentWorkouts}
+          </ul>
+        
+      </div>
     </div>
   )
 }
